+++ metaeditor_safetensors/resources/editor_panel.ui
@@ -291,7 +291,7 @@
   <customwidget>
    <class>ImageWidget</class>
    <extends>QGraphicsView</extends>
-   <header>metaeditor_safetensors.ui.image_widget</header>
+   <header>metaeditor_safetensors.widgets.image_widget</header>
   </customwidget>
  </customwidgets>
  <resources/>
