+++ metaeditor_safetensors/resources/icon.png
