+++ metaeditor_safetensors/resources/resources.qrc
