+++ metaeditor_safetensors/resources/resources_rc.py
