+++ metaeditor_safetensors/resources/style.qss
