+++ metaeditor_safetensors/widgets/image_widget.py
